Updated Javadoc and README
