+++ src/main/java/gobalconstants/constants.java
@@ -1,8 +1,12 @@
 package gobalconstants;
 
+/**
+ * Class to handle Constants
+ *
+ * @author Pabitra Swain (contact.the.sdet@gmail.com)
+ */
 public class constants {
     public static final String EMPTY_STRING = "";
-
     public static final String APP_DATE_FORMAT = "dd MMM EEE, yyyy";
     public static final String PREFERRED_DATE_FORMAT = "dd MMM yyyy";
 }
