+++ src/main/java/pages/pageobjects/MenuPage.java
@@ -4,6 +4,12 @@ import io.appium.java_client.AppiumDriver;
 
 import java.util.List;
 
+/**
+ * Page Object class for Menu Drawer page to have modular methods to handle element interactions
+ *
+ * @author Pabitra Swain (contact.the.sdet@gmail.com)
+ */
+@SuppressWarnings("unused")
 public class MenuPage extends HomePage {
     /**
      * Constructor to initialize AppiumUtils.
@@ -15,27 +21,62 @@ public class MenuPage extends HomePage {
         super(driver);
     }
 
+    /**
+     * Click on the hamburger menu icon.
+     *
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public void clickOnHamburgerMenu() {
         click(homePage.getDrawerButton());
     }
 
-
+    /**
+     * Check if the menu drawer is displayed.
+     *
+     * @return true if the menu drawer is displayed, false otherwise.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public boolean isMenuDrawerDisplayed() {
         return isVisible(homePage.getMenuDrawer());
     }
 
+    /**
+     * Close the menu drawer by swiping left.
+     *
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public void closeMenuDrawer() {
         swipeLeft(homePage.getMenuDrawer());
     }
 
+    /**
+     * Check if the login/sign-up button is displayed in the menu drawer.
+     *
+     * @return true if the login/sign-up button is displayed, false otherwise.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public boolean isLoginSignUpButtonDisplayed() {
         return isVisible(homePage.getLoginSignUpButton());
     }
 
+    /**
+     * Get the text of primary items in the menu drawer.
+     *
+     * @return a list containing the text of primary items in the menu drawer.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public List<String> drawerPrimaryItems() {
         return getElementsTextContent(homePage.getPrimaryItemsInMenuDrawer(), true);
     }
 
+    /**
+     * Check if a section is displayed on the menu drawer.
+     *
+     * @param section The name of the section to check.
+     * @return true if the section is displayed, false otherwise.
+     * @throws IllegalArgumentException if the section name is invalid.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public boolean isSectionDisplayedOnMenuDrawer(String section) {
         switch (section.toLowerCase()) {
             case "my trips":
@@ -49,6 +90,14 @@ public class MenuPage extends HomePage {
         }
     }
 
+    /**
+     * Get the text of items under a section in the menu drawer.
+     *
+     * @param sectionName The name of the section.
+     * @return a list containing the text of items under the section.
+     * @throws IllegalArgumentException if the section name is invalid.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public List<String> getItemsUnderSection(String sectionName) {
         switch (sectionName.toLowerCase()) {
             case "my trips":
@@ -62,6 +111,14 @@ public class MenuPage extends HomePage {
         }
     }
 
+    /**
+     * Verify if a specific item is displayed at the bottom of the menu drawer.
+     *
+     * @param item The name of the item to verify.
+     * @return true if the item is displayed, false otherwise.
+     * @throws IllegalArgumentException if the item name is invalid.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public boolean verifyMenuDrawerBottom(String item) {
         switch (item.toLowerCase()) {
             case "rate us link":
@@ -75,6 +132,12 @@ public class MenuPage extends HomePage {
         }
     }
 
+    /**
+     * Get the app version displayed at the bottom of the menu drawer.
+     *
+     * @return the app version.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public String getAppVersion() {
         return getElement(homePage.getMenuDrawerBottomLinks("App Version"))
                 .getText().split("Version ")[1];
