+++ src/main/java/utils/CommonUtils.java
@@ -2,57 +2,62 @@ package utils;
 
 import io.cucumber.java.Scenario;
 import logger.Log;
-import org.openqa.selenium.By;
-import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.Properties;
 
 import static engine.Engine.getDriver;
 import static io.github.the_sdet.cucumber.CucumberUtils.*;
-import static io.github.the_sdet.cucumber.CucumberUtils.attachScreenshot;
 
+/**
+ * Class to handle common utilities
+ *
+ * @author Pabitra Swain (contact.the.sdet@gmail.com)
+ */
 public class CommonUtils {
-    public static WebElement findElementByMultipleLocators(WebDriver driver, String... xPaths) {
-        for (String xpath : xPaths) {
-            try {
-                return driver.findElement(By.xpath(xpath));
-            } catch (org.openqa.selenium.NoSuchElementException e) {
-                Log.info("No element found for Xpath: " + xpath);
-            }
-        }
-        throw new NoSuchElementException("Element NOT found for any of the provided locators...");
-    }
-
     /**
-     * This method retrieves and returns the feature name of which the scenario is part of
+     * Extracts the feature name from the given Scenario object.
      *
-     * @param scenario Cucumber scenario name
-     * @return Cucumber feature name from scenario
+     * @param scenario        The Scenario object from Cucumber.
+     * @param withPackageName A boolean indicating whether to include package name along with feature name.
+     * @return The feature name extracted from the Scenario.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
      */
     public static String getFeatureNameFromScenario(Scenario scenario, boolean withPackageName) {
         String[] test = scenario.getUri().toString().split("/");
         String featureName = test[1].split("\\.")[0];
         String packageName = test[0].split(":")[1];
-        Log.info("Feature: "+featureName);
-        Log.info("Package: "+packageName);
+        Log.info("Feature: " + featureName);
+        Log.info("Package: " + packageName);
         if (withPackageName)
             return packageName + " - " + featureName;
         else return featureName;
     }
 
+    /**
+     * Enum representing the status of a test scenario.
+     *
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
     public enum STATUS {
-        PASS ("PASS"), FAIL ("FAIL");
+        PASS("PASS"), FAIL("FAIL");
         public final String status;
-        STATUS(String status){
+
+        STATUS(String status) {
             this.status = status;
         }
     }
 
-    public static void attachScreenshotPerConfig(Scenario scenario){
+    /**
+     * Attaches screenshot to the Cucumber report based on the configuration.
+     *
+     * @param scenario The Scenario object from Cucumber.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
+    public static void attachScreenshotPerConfig(Scenario scenario) {
         Properties properties = ConfigReader.getProperties();
         if (properties.getProperty("screenshot").equals("only.fail")) {
             if (scenario.isFailed()) {
@@ -67,6 +72,27 @@ public class CommonUtils {
         } else {
             Log.info("Taking Screenshot.");
             attachScreenshot(getDriver());
+        }
+    }
+
+    /**
+     * Format a date string from one format to another.
+     *
+     * @param inputDateStr     The input date string.
+     * @param inputDateFormat  The input date format.
+     * @param outputDateFormat The output date format.
+     * @return the formatted date string.
+     * @author Pabitra Swain (contact.the.sdet@gmail.com)
+     */
+    public static String formatDate(String inputDateStr, String inputDateFormat, String outputDateFormat) {
+        DateFormat inputFormatter = new SimpleDateFormat(inputDateFormat);
+        DateFormat outputFormatter = new SimpleDateFormat(outputDateFormat);
+        try {
+            Date inputDate = inputFormatter.parse(inputDateStr);
+            return outputFormatter.format(inputDate);
+        } catch (ParseException e) {
+            Log.error("Parse Exception...", e);
+            return null;
         }
     }
 }
