+++ src/test/java/runner/TestRunner.java
@@ -3,7 +3,11 @@ package runner;
 import org.junit.platform.suite.api.*;
 
 import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
-
+/**
+ * Runner class to glue features and step definitions
+ *
+ * @author Pabitra Swain (contact.the.sdet@gmail.com)
+ */
 @Suite
 @IncludeEngines("cucumber")
 @SelectClasspathResource("features")
